♻️ Extract account-deletion feature to break the user to billing import cycle
The deletion lifecycle (schedule/cancel actions, reaper, summary, banner)
orchestrates across user, billing, and PostHog. Living in features/user it
created a user to billing edge that merged the grandfathered billing/space
cycle with user and licensing into one large SCC. As its own feature nothing
imports it except app/, so the cycle disappears. hardDeleteUser stays in
features/user (pure user-row concern, shared with the admin delete action).
