+++ apps/web/src/features/account-deletion/components/account-deletion-banner.tsx
@@ -1,4 +1,4 @@
-import { getScheduledDeletionDate } from "@/features/user/utils";
+import { getScheduledDeletionDate } from "@/features/account-deletion/utils";
 import { Trans } from "@/i18n/client";
 import { getSession } from "@/lib/auth";
 import { formatDateTime } from "@/lib/datetime/format";
